+++ cmd/login/runner_test.go
@@ -165,6 +165,26 @@ func TestLogin(t *testing.T) {
 			startConfig: createValidTestConfig("", true),
 			expectError: unknownUrlError,
 		},
+		// Logging into non default context
+		{
+			name:        "case 12",
+			startConfig: createNonDefaultTestConfig(),
+			mcArg:       []string{"arbitraryname"},
+			flags: &flag{
+				WCCertTTL: "8h",
+			},
+			expectError: contextDoesNotExistError,
+		},
+		// Logging into non default context
+		{
+			name:        "case 13",
+			startConfig: createNonDefaultTestConfig(),
+			mcArg:       []string{"arbitraryname"},
+			flags: &flag{
+				WCCertTTL:      "8h",
+				EnforceContext: true,
+			},
+		},
 	}
 
 	for i, tc := range testCases {
@@ -407,6 +427,29 @@ func createValidTestConfig(wcSuffix string, authProvider bool) *clientcmdapi.Con
 		config.AuthInfos["gs-user-"+clustername] = &clientcmdapi.AuthInfo{
 			Token: token,
 		}
+	}
+
+	return config
+}
+
+func createNonDefaultTestConfig() *clientcmdapi.Config {
+	const (
+		server      = "https://anything.com:8080"
+		token       = "the-token"
+		clustername = "arbitraryname"
+	)
+
+	config := clientcmdapi.NewConfig()
+	config.Clusters[clustername] = &clientcmdapi.Cluster{
+		Server: server,
+	}
+	config.Contexts[clustername] = &clientcmdapi.Context{
+		Cluster:  clustername,
+		AuthInfo: "user-" + clustername,
+	}
+	config.CurrentContext = clustername
+	config.AuthInfos["user-"+clustername] = &clientcmdapi.AuthInfo{
+		Token: token,
 	}
 
 	return config